% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_analysis.R
\name{cluster_formula}
\alias{cluster_formula}
\title{cluster_formula: gives the formula for clustering standard errors based on what fixed effects variables you tell it to keep, for use in the argument of modelsummary}
\usage{
cluster_formula(reg_eq = NULL, fe_vars = NULL, verbose = TRUE)
}
\arguments{
\item{reg_eq}{additionally, a regression equation of the format that the fixest package will use, which is to say of the type y ~ x | year + country. In this case, fixest_cluster_formula will pick out that the fe_vars are c("year","country") so that modelsummary or feols can cluster appropriately}

\item{fe_vars}{default NULL, a character vector of the fixed effects you're using}

\item{verbose}{default = TRUE, will tell you if it thinks you inputted something wrong and what the defaults are}
}
\value{
a formula with the clustered variables that you can include in the vcov argument of modelsummary to get you clustered standard errors. If it thinks you've inputted a regression equation without fixed effects, it's going to try to default to heteroskedasticity robust standard errors with "HC1"
}
\description{
cluster_formula: gives the formula for clustering standard errors based on what fixed effects variables you tell it to keep, for use in the argument of modelsummary
}
