% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_download.R
\name{collect_filenames_urls}
\alias{collect_filenames_urls}
\title{Collect imported urls and filenames into a list of data frames for easy calling}
\usage{
collect_filenames_urls(
  df_name_vec,
  sheetnames_vec,
  filenames_vecname_vec,
  url_vecnames_vec,
  ...
)
}
\arguments{
\item{df_name_vec}{vector of names for the dfs}

\item{sheetnames_vec}{vector of names of the sheet to pull from}

\item{filenames_vecname_vec}{vector of the filenames vector's names}

\item{url_vecnames_vec}{vector of the url vector's names}
}
\description{
Collect imported urls and filenames into a list of data frames for easy calling
}
