% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_startup_palette.R
\name{display_hex_colors}
\alias{display_hex_colors}
\title{Display Hex Colors}
\usage{
display_hex_colors()
}
\description{
Puts into the console HEX combinations that you can define
}
\examples{
display_hex_colors()
yale_lblue <- "#63aaff"
yale_medblue   <- "#286dc0"
yale_blue   <- "#00356b"
woodsy_green <- "#228B22"
}
