% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_download.R
\name{download_data}
\alias{download_data}
\title{Function for downloading single data file from a website, zip or password protected file}
\usage{
download_data(
  data_subfolder = NULL,
  data_raw = NULL,
  filename = NULL,
  url,
  zip_file = FALSE,
  pass_protected = FALSE,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{data_subfolder}{name of subfolder in the 01_raw data folder to place data in. Null defaults to just living in the data_raw folder}

\item{data_raw}{location (file path) of raw data folder. Defaults to here("data","01_raw")}

\item{filename}{if not a zip file and need to use download.file, needs to provide a name for the ultimate file}

\item{url}{the url of the folder to download}

\item{zip_file}{TRUE or FALSE, if TRUE use file path of a zip folder}

\item{pass_protected}{TRUE or FALSE, if TRUE you'll get prompted for username and password. if true you need to provide username = and password =}
}
\value{
data file in the particular subfolder
}
\description{
Function for downloading single data file from a website, zip or password protected file
}
