% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{get_in_out_intersections}
\alias{get_in_out_intersections}
\title{get intersections in one group but not another}
\usage{
get_in_out_intersections(sf_of_interest, yes_sf, no_sf)
}
\arguments{
\item{sf_of_interest}{the sf that we want to get that intersects yes_sf but not no_sf}

\item{yes_sf}{the sf we want to intersect}

\item{no_sf}{the sf we want to remove the intersections from}
}
\description{
get intersections in one group but not another
}
