% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_plots.R
\name{ggsave_plot}
\alias{ggsave_plot}
\title{Wraps around ggsave with default settings useful for a basic figure}
\usage{
ggsave_plot(
  output_folder = NULL,
  plotname,
  filename,
  width = 9,
  height = 5,
  dpi = 300
)
}
\arguments{
\item{output_folder}{the path to the output directory. defaults to here::here("output","03_maps")}

\item{plotname}{the name of the plot that you generated}

\item{filename}{the name of the output filename, e.g. "my-plot.png"}

\item{width}{numeric, defaults to 9 (inches)}

\item{height}{numeric, height, defaults to 5 (inches)}

\item{dpi}{pixel count, defaults to 300}
}
\value{
a saved png in the output folder
}
\description{
Wraps around ggsave with default settings useful for a basic figure
}
