% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_startup_download-functions.R
\name{import_filenames_urls}
\alias{import_filenames_urls}
\title{Import urls and filenames from an excel sheet}
\usage{
import_filenames_urls(
  names_path,
  sheetname,
  filenames_index,
  url_index,
  filenames_vecname,
  url_vecname
)
}
\arguments{
\item{names_path}{path leading to the excel with names}

\item{sheetname}{name of the sheet to pull from}

\item{filenames_index}{index of the column where filenames will be}

\item{url_index}{index of the col where urls will be}

\item{filenames_vecname}{name you want for the vector}

\item{url_vecname}{name for the url vec}
}
\description{
Import urls and filenames from an excel sheet
}
