% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_startup_analysis-functions.R
\name{lm_model}
\alias{lm_model}
\title{Function for outputting an LM model}
\usage{
lm_model(df, outvar, covars, ...)
}
\arguments{
\item{df}{data frame the vars come from}

\item{outvar}{string, output var}

\item{covars}{of form c("covar1", "covar2", etc) all covariates}
}
\value{
lm model
}
\description{
Function for outputting an LM model
}
