% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_analysis.R
\name{modelsummary_reg_default}
\alias{modelsummary_reg_default}
\title{modelsummary_reg_default}
\usage{
modelsummary_reg_default(
  reg_vars_list,
  data,
  my_title = "My regression table \\\\label{tab:reg-table}",
  table_notes = c("Robust standard errors given in parentheses."),
  cov_labels = FALSE,
  fe_names = NULL,
  depvar_means = NULL,
  stars = FALSE,
  gof_omit = "AIC|BIC|RMSE|Log.Lik|Std.Errors|FE:|Adj.|F",
  fmt = 4,
  out_path = here::here("output", "01_tables"),
  output_filename = "regression_table",
  export_output = TRUE,
  print_models = TRUE,
  ...
)
}
\arguments{
\item{reg_vars_list}{requires a list where each element of the list is also a list.}

\item{my_title}{title for your table}

\item{table_notes}{notes to put in your table}

\item{cov_labels}{labels for the covariates. if you input this you need to make sure it's the right length for the covariates you're including}

\item{fe_names}{defaults to NULL, in which case we take the column names of the variables inputted as FE vars. could instead give a character vector e.g. c("Year FE","Country FE") to put into the modelsummary output. You need to make sure that this character vector is in the order that this function will pick out, though, so I recommend running it through first with fe_names = NULL to see that the output accords with what you think, and then adjusting.}

\item{depvar_means}{default of NULL takes the mean of the dependent variable for each column, rounded to 2 significant digits, with missing values removed before the mean is calculated. If you want something different, input a numeric vector of the means that you want for each column of your regression. If not NULL, this vector therefore needs to have the same number of elements as reg_vars_list does.}

\item{stars}{defaults to FALSE, from modelsummary}

\item{gof_omit}{goodness of fit statistics to ignore, defaults to keeping mostly the R2 and Adj R2}

\item{fmt}{format of the final table, defaults to 4 decimal points, from modelsummary}

\item{out_path}{defaults to here::here("output","01_tables"), the place to send table output if you're saving it. will create the folder if it doesn't exist already}

\item{output_filename}{character vector, name for output filename. defaults to "regression_table"}

\item{export_output}{defaults to TRUE, in which case it saves a .docx and a .tex file into the file path given by file.path(out_path,paste0(output_filename,".tex")) and file.path(out_path,paste0(output_filename,".docx"))}

\item{print_models}{default is TRUE, whether you want to print the list of models with output to the console}

\item{...}{additional options to put into modelsummary}

\item{format}{defaults to "latex", in modelsummary}

\item{escape}{defaults to FALSE, in modelsummary so that output is less buggy}
}
\value{
a modelsummary object
}
\description{
modelsummary_reg_default takes in a list of regression variables formatted in a particular way and a data frame, and outputs a modelsummary table with reasonable defaults for these regressions. It currently supports using fixest::feols with fixed effects (in which case it clusters standard errors by the fixed effects categories); as well as lm() for regression equations which do not have fixed effects variables inputted. It tries to cleverly add in dependent variable means and fixed effect rows.
See the vignette \href{https://stallman-j.github.io/ekonomR/vignettes/fixed-effects-estimation}{Fixed Effects Estimation} for how this function comes together and a couple examples of use.
}
\examples{
data(ghg_pop_gdp)
reg_1_vars <- list(outvar = "gcb_ghg_territorial_pc",
regvars = c("gdp000_pc","I(gdp000_pc^2)","I(gdp000_pc^3)"),
fevars  = NULL)
reg_2_vars <- list(outvar = "gcb_ghg_territorial_pc",
                   regvars = c("gdp000_pc","I(gdp000_pc^2)","I(gdp000_pc^3)"),
                   fevars = "year"))
reg_vars_list <- list(reg_1_vars,reg_2_vars)
# default output
modelsummary_reg_default(reg_vars_list, data = ghg_pop_gdp)
# save and adjust ex post, so don't export
my_table <- modelsummary_reg_default(reg_vars_list, data = ghg_pop_gdp, export_output = FALSE)
my_table <- my_table \%>\% tinytable::group_tt(j = list("GHGpc" =2:3))
if (!dir.exists(here::here("output","01_tables"))) dir.create(here::here("output","01_tables"), recursive = TRUE)
tinytable::save_tt(my_table, output = here::here("output","01_tables","fixed_effects_table.tex"), overwrite = TRUE)

}
