% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{negative_buffer_ring}
\alias{negative_buffer_ring}
\title{Get Negative Buffer Ring}
\usage{
negative_buffer_ring(sf, size)
}
\arguments{
\item{sf}{original sf you want negative buffer for}

\item{size}{size of buffer, positive value}

\item{neg_buffer_if_possible:}{take the negative buffer if it exists, otherwise just use the original}
}
\description{
Gets a buffer
}
