% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_startup_spatial-functions.R
\name{positive_buffer_ring}
\alias{positive_buffer_ring}
\title{positive_buffer}
\usage{
positive_buffer_ring(sf, size)
}
\arguments{
\item{sf}{original sf you want negative buffer for}

\item{size}{size of buffer, positive value}
}
\description{
gets a buffer
}
