% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_extract_to_panel.R
\name{raster_extract_to_panel}
\alias{raster_extract_to_panel}
\title{Extracts terra rasters to a vector SF}
\usage{
raster_extract_to_panel(
  terra_raster,
  vector_sf,
  cast_vector_sf = TRUE,
  save_raster_copy = FALSE,
  raster_out_path = getwd(),
  raster_out_filename = "terra_raster.rds",
  extracted_out_path = getwd(),
  extracted_out_filename = "terra_raster_extracted_vector_cast.rds",
  vector_cast_out_path = getwd(),
  vector_cast_out_filename = "vector_cast.rds",
  layer_substrings,
  func = "mean",
  weights = NULL,
  time_interval = "months",
  remove_files = FALSE
)
}
\arguments{
\item{terra_raster}{a terra raster, currently required to be from 0-360 degrees longitude and -90 to 90 degrees latitude,
and must have a time dimension given}

\item{vector_sf}{an sf object with polygons, points or lines to be extracted to}

\item{save_raster_copy}{default TRUE, logical if you want to save a copy of the raster you bring in
with possibly updated coordinate reference system or rotated correctly}

\item{raster_out_path}{file path to save a copy of the raster to}

\item{raster_out_filename}{if saving a raster copy, name of the file}

\item{extracted_out_path}{filepath to where we output the data of extracting raster to vector}

\item{extracted_out_filename}{filename of output file from extracted raster to vector (no geometry kept)}

\item{vector_cast_out_path}{filepath to where the vector sf that gets casted is output}

\item{vector_cast_out_filename}{filename for the vector sf output}

\item{layer_substrings}{character vector, giving the substrings which identify what layers
assumes each element of this will give layers which generate the same longitudinal time sequence}

\item{func}{character vector, default "weighted_sum," function to be used for aggregating}

\item{weights}{character vector, default "area" to be used for weighting the function}
}
\description{
\code{raster_extract_to_panel} returns polygon shapefiles with longitudinal data from rasters
extracted to those polygons and writes two sf files, one with the polygons shapefiles, which are
cast so that no single row contains multiple geometries,
and one with a long data frame that includes the raster extracted values
and an identifier that allows for linking to the sf polygons.
}
\author{
jillianstallman
}
