% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_analysis.R
\name{reg_equation}
\alias{reg_equation}
\title{Get Regression Equation}
\usage{
reg_equation(
  outcome_var = "lge_15",
  regressor_vars = c("gdp_pc", "tfr"),
  fe_vars = NULL
)
}
\arguments{
\item{outcome_var}{a string vector with the outcome variable name}

\item{regressor_vars}{a character vector with all the regressors (not fixed effects) variable names}

\item{fe_vars}{if fixed effects desired, a character vector of the variable names to take fixed effects of.}
}
\value{
A regression formula that can either be used as an input to the lm function or fixest::feols if fe_vars is not null
}
\description{
Get Regression Equation
}
