% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_plots.R
\name{theme_minimal_plot}
\alias{theme_minimal_plot}
\title{Plot Theme Defaults}
\usage{
theme_minimal_plot(
  legend_text_size = 8,
  legend_title_size = 10,
  legend_position = c(0.2, 0.3),
  axis_title_x = element_text(color = "black"),
  axis_title_y = element_text(color = "black"),
  axis_text_x = element_text(color = "darkgrey"),
  axis_text_y = element_text(color = "darkgrey"),
  ...
)
}
\arguments{
\item{legend_text_size}{replaces legend.text in theme() with legend.text = element_text(linewidth = legend_text_size. defaults to 8.)}

\item{legend_title_size}{replaces legend.title = element_text(linewidth = legend_title_size), defaulting to 10}

\item{legend_position}{replaces legend.position, using c(a,b) where the a and b are numbers in 0 to 1. a gives the left-right movement (.2 is mostly on the left) and b gives the up-down location. use legend_position = "none" for no legent.}

\item{axis_title_x}{defaults to making axis title black. use axis_title_x = element_blank() to remove}

\item{axis_title_y}{defaults to making axis title black. use axis_title_y = element_blank() to remove}

\item{axis_text_x}{defaults to making the axis text a dark grey. use element_blank() to remove}

\item{axis_text_y}{defaults to making axis text dark gray. use element_blank() to remove}

\item{...}{additional options for ggplot2::theme()}
}
\value{
wrapper around ggplot2::theme_minimal() and theme()
}
\description{
Wraps around ggplot2's theme_minimal with default settings that are nice and professional for basic plotting
}
