% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sf_wf.R
\name{write_sf_wf}
\alias{write_sf_wf}
\title{write_sf_wf}
\usage{
write_sf_wf(
  my_sf,
  out_path = here::here(),
  out_name = "my_out_sf",
  out_type = c(".rds", ".shp")
)
}
\arguments{
\item{my_sf}{the input sf, the one which will have rows kept and saved as an output}

\item{out_path}{file path, place to save the output. gets created if it doesn't already exist}

\item{out_name}{filename of basin to use}

\item{out_type}{default c(".rds",".gpkg") saves a RDS file and a .gpkg file. options also include ".shp", ".geojson". If you give an unsupported file type, it won't save}
}
\value{
the intersected sf (i.e. the rows of \code{my_sf} which had a positive intersection with \code{intersecting_sf})
}
\description{
\code{write_sf_wf} Wrapper around sf's \code{write_sf}that, for a given \code{my_sf}, saves it in a particular file path. the \verb{_wf} is for "workflow" because this is a common workflow
}
\author{
jstallman
}
