% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_rent_proposal.R
\name{calculate_rent_proposal}
\alias{calculate_rent_proposal}
\title{Calculate rent Proposal}
\usage{
calculate_rent_proposal(
  bls_xlsx_file,
  start_month = "Jun",
  start_year = "2021",
  end_month = "Mar",
  end_year = "2025",
  start_rent = 1500,
  skip_val = 10,
  ...
)
}
\arguments{
\item{bls_xlsx_file}{full filepath to the xlsx file downloaded from the BLS}

\item{start_month}{character, three-letter month as given in BLS. Default "Jun"}

\item{start_year}{character, year of start date. Default "2021"}

\item{end_month}{character, three-letter month as given by BLS, e.g. "Mar"}

\item{end_year}{character, year of end date. Default is "2025"}

\item{start_rent}{numeric, starting rent. Default is 1500}

\item{skip_val}{numeric, number of rows to skip in the BLS xlsx data. Default 10 worked for the ones I downloaded}

\item{...}{additional parameters to put into the readxl::read_xlsx() call for the BLS file}
}
\value{
suggested proposed rent based on the inflation calculated.
}
\description{
\code{calculate_rent_proposal} takes in a BLS file and some start and end dates, then a starting rent, and suggests a proposed rent based on the increase in CPI from the start date to ending date.
}
\author{
jstallman
}
