% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_startup_download-functions.R
\name{download_data}
\alias{download_data}
\title{Function for downloading data from a website, zip or password protected file}
\usage{
download_data(
  data_subfolder,
  data_raw = data_raw,
  filename = NULL,
  url,
  zip_file = FALSE,
  pass_protected = FALSE,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{data_subfolder}{name of subfolder in the 01_raw data folder to place data in}

\item{data_raw}{location (file path) of raw data folder}

\item{filename}{if not a zip file and need to use download.file, needs to provide a name for the ultimate file}

\item{url}{the url of the folder to download}

\item{zip_file}{TRUE or FALSE, if TRUE use file path of a zip folder}

\item{pass_protected}{TRUE or FALSE, if TRUE you'll get prompted for username and password. if true you need to provide username = and password =}
}
\description{
Function for downloading data from a website, zip or password protected file
}
