% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_startup_download-functions.R
\name{download_multiple_files}
\alias{download_multiple_files}
\title{Download Files}
\usage{
download_multiple_files(
  data_subfolder,
  data_raw,
  base_url,
  sub_urls,
  filenames,
  pass_protected = FALSE,
  zip_file = FALSE,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{data_subfolder}{name of subfolder in the 01_raw data folder to place data in}

\item{data_raw}{location (file path) of raw data folder}

\item{base_url}{the url of the folder to download}

\item{sub_urls}{the additional endings to the base url that you'd like to download separately, a vector of characters}

\item{filenames}{the names of files, without paths, to attach to those suburls.}

\item{create_folder}{if TRUE, create a subfolder}
}
\description{
Function for downloading multiple data files from a website
}
