% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{lm_model}
\alias{lm_model}
\title{Function for outputting a linear model}
\usage{
lm_model(df, outvar, covars, ...)
}
\arguments{
\item{df}{data frame the vars come from}

\item{outvar}{string, output var}

\item{covars}{of form c("covar1", "covar2", etc) all covariates}
}
\value{
lm model
}
\description{
Outputs a linear model from an outvar and covariates within a data frame, see https://gis.stackexchange.com/questions/403811/linear-regression-analysis-in-r
}
