% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_analysis.R
\name{lm_robust_se}
\alias{lm_robust_se}
\title{Function for getting robust ses, requires sandwich package}
\usage{
lm_robust_se(lm_model, type = "HC1")
}
\arguments{
\item{lm_model}{the lm model, output of lm_model or just a function arrived at via lm()}

\item{type}{type of SEs, for stata robust it's "HC1"}
}
\value{
lm_robust_se robust SEs
}
\description{
Function for getting robust ses, requires sandwich package
}
